Probability of Default of loan accounts
The credit risk departments of banks mainly deal with the risk of account customers’ default on loans. The primary mission of F-IRB (Foundation Internal Ratings Based Approach) of credit risk departments is to provide the PD (Probability of Default) internally, associated with the LGD (Loss Given Default) and EAD (Exposure At Default) to calculate the EL (Expected Losses) to determine the Minimum Capital Requirement. 
Usually, the amount of minimum capital requirement involves the profitability of the bank’s operation. This task is to provide the PD internally based on the banks’ retail data. This gives the bank the ability to reserve sufficient money to invest without affecting the ability to resist the losses caused by defaults.